Handle kwargs in method calls
If a user passes in kwargs which are not understood by PJRmi then we pass them
on to method being called. This allows for easier creation of duck-typed classes
which need to handle kwargs.

As part of this we also add a way for dynamic binding to ignore certain
methods which might otherwise cause confusion when we are trying to bind
kwargs. It could be handy in other circumstances too.

This changes the wire format and so includes a bump in the minor version
number.