Prevent accidental addition of attributes
We create a guard against the accidental creation of attributes on Java object
instances this will help catch typos when folks are trying to assign values to
Java object instance members from Python.

As part of this commit we also rename a few attributes so that they are
better namespeced.