Make Python str interpretation work as expected
We make String be a more specialised type than char[] and byte[]. This
is because a Python str can be interpreted as all three and this leads
to binding ambiguity in methods which take all three (like the String
CTOR and valueOf() methods). If we pass in a Python str then we should
generally expect that to bind as a String.

Similarly, a char[] is more specific than a byte[] so that str instances
are turned into char[]s in preference over byte[]s.