Adding class redefinition support
Allow users to easily rewrite classes on the fly.

This adds a new directive to the protocol which technically would mean a
bump in the minor version number. However, since this is not really a
directive which we expect folks to use to much, and one which will not
cause undue fallout should its use be attempted, we only bump the
sub-minor number.