Avoid hopping through memory if possible
Most cubes may well be smaller that 2^30 in size and, if so, we should
avoid the extra hop via the array-of-arrays. Also, since we know that
the arrays will never change under our feet, we don't need to be using
an AtomicReferenceArray to access the sub-arrays.

Avoiding the extra memory hop yields a 30% speed-up for some access
patterns.