Don't clobber Java's versions of Python methods
If the Java side has created its own version of a Java method then we are now
careful not to overwrite that with the PJRmi implementation.

For example, should a Java Iterable want to have a special __iter__() method
(which it knows will be called by the Python side): we let it do so, in favour
of imposing the PJRmi version of the method, which wraps the Java iterator()
code.